pyproject.toml
[tool.poetry]
name = "datasciencematerial"
version = "0.1.0"
description = "content related to datascience"
authors = ["udhay Prakash <uday3prakash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
numpy = "^1.25.0"
pandas = "^2.0.2"
matplotlib = "^3.7.1"
openpyxl = "^3.1.2"
scikit-learn = "^1.2.2"
yfinance = "^0.2.20"
wordcloud = "^1.9.2"
pypdf2 = "^3.0.1"
nltk = "^3.8.1"
python-docx = "^0.8.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
